requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.2.0
python-multipart
opencv-python-headless
numpy
tensorflow-cpu>=2.14.0
google-generativeai
pydantic>=2.0
pydantic-settings
Pillow
